Add osv_execve, osv_waittid
osv_execve is replacement for execve, it starts given program in a new
thread, and in a new ELF namespace. A thread id is returned, so that caller
can later wait on thread termination - osv_waittid is used for that.
Optionally, caller can be notified of thread termination via file
descriptor returned by eventfd().

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <1457457195-17597-2-git-send-email-justin.cinkelj@xlab.si>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>