tests: another mutex benchmark
This test adds a paramterized mutex benchmark, tests/tst-mutex2.so.
This benchmark simulates a situation where a mutex is held for very short
durations, where we could achieve better performance if the mutex retried
the lock for a short while ("spinning") before putting the thread to sleep.

The goal of this test is to reproduce a situation discovered by Justin Cinkelj
in the sysbench benchmark (sysbench --test=memory --num-threads=4 run).
That benchmark has some intensive memory-accessing code, but it very often
(as much as a million times a second) updates a counter protected by a mutex.
That sysbench benchmark runs significantly slower on OSv than it does on Linux,
and we see with freq.py hundreds of thousands of context switches per second
caused by sleeping mutex.

This benchmark reproduces the sysbench slowdown with a much simpler benchmark.
We run N threads, each iteratively increments a mutex-protected counter and
then does a small amount of computational work.

Some example runs of this benchmark on one particular 4-core machine, followed
by analysis of these results:

 +--------------------+-----------------------+----------------------+
 | Run parameters     |  On OSv guest (op/s)  | On Linux host (op/s) |
 +--------------------+-----------------------+----------------------+
 | 1 0                |  37,000,000           | 49,000,000           |
 | 1 500              |  5,700,000            | 6,100,000            |
 | 2 500              |  3,700,000            | 9,300,000            |
 +--------------------+-----------------------+----------------------+

The first two lines show us that when uncontended, OSv's mutex lock and unlock
is a bit slower than those on Linux. This is not something to rejoice about,
but not the big problem... The big problem is the third line, where OSv's
results are significantly worse than those on Linux (exactly as noticed
in the sysbench benchmark). Moreover, the OSv results have very large variance.

We would expect the results of "2 500" to be *better* than those of 1 500
(as indeed happened on Linux but not on OSv), because this benchmark does
quite a bit of computation outside the mutex (a loop of 500 steps on every
iteration), so if we do it on two threads, we can get more work done. But in
OSv, adding the second thread actually slowed down the benchmark - when one of
the threads finds a locked mutex it goes to sleep (which takes a substantial
amount of time) and also forces the other thread thread to wake it once it
released the mutex, also taking time. On the other hand, Linux's mutex
tries to spin for a bit before going to sleep, and usually the lock is
gained after only short time of spinning, without a context switch occuring.

The reason for the high variance in OSv's results in the "2 500" case is
as follows: If OSv places the two threads on the same CPU, the result should
be identical to that of one thread (5,700,000). But if OSv places the two
threads on different CPUs, the results are significantly worse (around
than 3,300,000). Now, because OSv moves the threads between CPUs rather
sporardically, we end up with a result which something between those two
numbers, and in each run we end up with a different number depending on
how much time the threads spent on the same CPU.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1460384096-11635-1-git-send-email-nyh@scylladb.com>