Make sure lzloader.elf has integer number of blocks
During x86-64 boot process, lzloader is read from disk block by block
using BIOS interrupt 13.

However, the end lzloader.elf is not aligned, so it does not have integer
number of blocks. It does not produce any issue in case of usr.img because zfs
partitions are stored after lzloader. But in the loader.img, the last
block of it is going to be truncated by the end of the disk.

Because of this int13 fails to process the last read request, and a tiny
part of OSv kernel is missing in the memory.

This patch just truncates lzloader.elf up to nearest integer number of
blocks.

Signed-off-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Message-Id: <1460632342-11202-3-git-send-email-iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>