Makefile: allow using a different bootfs.manifest
Bootfs.manifest is generated anyways by the module.py script.
Using generated instead of hardcoded opens the possibility to alternate
the context of resulting bootfs image.

However, bootfs.manifest is generated from scripts/build, so if user
wants to build kernel only, the bootfs.manifest is not exists at that
moment. This solved by extra parameter to the makefile:
bootfs_manifest=<path_to_the_manifest>

If no parameter specified, old good bootfs.manifest.skel is used.

This functionality is not used in this patch yet. It is preparation for
the following bootfs-only configuration.

Signed-off-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Message-Id: <1460650745-17348-1-git-send-email-iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>