scripts/build: fix triple run of make
Noticed and originally fixed by Yuri Volchkov:

 "In case of compatibility hack (make image=<any>), the MAKEFLAGS passed
  to the scripts/build might or might not have the space before "image=".
  In the second case, construction
  MAKEFLAGS=${MAKEFLAGS/ image=/ zzz=}
  will not work. This will cause the extra run of the make."

In this patch, in addition to the rule replacing "image=" preceded by
a space, we add a second rule replacing "image=" in the beginning of the
line. We want to avoid replacing the string "image=" in the middle of a
word (not that this is likely, but let's play it safe...).

Suggested-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>