Makefile: clean up bootfs.o handling
One of the dependencies we had for loader.elf is bootfs.bin. But we
don't need bootfs.bin directly - we rather need bootfs.o (which in turn
depends on bootfs.bin). So we need *that* as a dependency for loader.elf,
not bootfs.bin. Replacing the bootfs.bin dependency also allows us to get
rid of the silly code to remove it from the link command :-)

Note that bootfs.o was already a dependency of loader.elf, because bootfs.o
was listed in $(objects). It shouldn't be, as bootfs.o is already (after
this patch) a separate dependency of loader.elf.

Finally, now that bootfs.o is removed from $(objects), it makes it easy
to compile $(objects) without building bootfs.bin.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>