System V IPC: don't destroy shared memory segment too early
As the shmctl() manual page explains, when shmctl(IPC_RMID) is called to
destroy a shared memory segment, the actual destruction should be delayed
until the last attachment (shmat()) is detached (with shmdt()).

Before this patch, this wasn't done properly - IPC_RMID destroyed the
segment immediately, including the internal file descriptor used to
represent it, and later calls to shmdt() would crash because the file
on a null pointer dereference (as the file descriptor which they assumed
exists, did not).

In this simple patch, shmat() uses dup() on the internal file descriptor,
to create another reference to the shared memory segment. IPC_RMID closes
only one duplicate, and shmdt() closes the duplicate previously created by
shmat(), so the end result is that the underlying object is only destoryed
once all the duplicates have been closed - i.e., both IPC_RMID was used and
all the attachments were detached.

This patch also adds a test for this case. This test crashed before this
patch and succeeds now.

This problem (of IPC_RMID destroying the segment too early) was noticed
when running OpenMPI with System V shared memory to communicate between
separate threads on the same VM.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1461143164-22959-1-git-send-email-nyh@scylladb.com>