Makefile: absolute path in (gcc|glibc|jdk|misc)base
These variables are calculated twice: in the Makefile and in the
scripts/build. In the Makefile all paths are relative, in the
scripts/build - absolute.

Since script mkbootfs is launched from Makefile, with working directory
$(out), it will search %(gccbase) at the wrong path.  This means
references from apps/<anyapp>/bootfs.manifest to any of these variables
would not work.

This patch makes this variables containing absolute pathes too.

Signed-off-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Message-Id: <1461274714-7230-2-git-send-email-iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>