mkbootfs.py: Allow an empty directory
For the purposes of ramfs-only images, we want to pass usr.manifest to
the scripts/mkbootfs.py. However user.manifest contains directories
records (/proc, /dev and /tmp).

To describe the directory in the bootfs metadata, a zero-size record is
created. The name in such record ends with a "/", so kernel can
recognize the directory without adding new fields into the
"struct bootfs_metadata".

Signed-off-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Message-Id: <1461274714-7230-3-git-send-email-iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>