elf: fix crash while unloading shared object
While working on fixing cloud-init compilation in Fedora 23 (issue #738)
I noticed a crash while cloud-init.so was being unloaded (before running
the actual application). It turns out we have the following sequence of
events:

1. Shared object A is run, and requires shared object B
2. A completes its run, reaches a reference count of 0, so it
   unloads (program::remove_object(A) is called)
3. As part of remove_object(), it clears A's "needed" list, so
4. Object B is no longer needed so we start to unload it too.
5. As part of unloading B, B's fini functions are called
6. One of B's fini functions calls a function never called before,
   causing object::resolve_pltgot to be called.
7. If resolve_pltgot finds that function in object A,
8. it uses A->shared_from_this() to save a reference to A.
   This crashes because A already reached reference count 0 (step 2 above).

Note that one of the conditions of this crash is the unusual situation
where B finds the functions it needs in A. This probably happened cloud-init
(issue #738) because cloud-init requires httpserver, but both contain some
identical code (like compiled templates).

This fix in this patch is to change step 7: We remove A from the module
list before running B's fini functions, rather than after in the existing
code.

Even after this patch, there remains the possibility of a different race
condition (which will need to be fixed separately): If one thread causes
a shared object to be unloaded, and another thread runs a new function
which is found in the object being unloaded, we can have a crash when
the object is unloaded between finding it, and saving a shared pointer to it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1462955300-18185-1-git-send-email-nyh@scylladb.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>