mlockall()/munlockall() - add stub warnings
Our mlockall()/munlockall() are stubs that do nothing. This would usually
be fine but still there are certain performance implications involved: Our
do-nothing implementation doesn't bring into memory an mmap()ed file nor
does it ensure it stay there, and similarly for the executable code itself.

So better warn (once) that this function was used, and was stubbed.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1463304374-15856-1-git-send-email-nyh@scylladb.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>