Makefile: overhaul "make image=..." compatibility hack
For quite a while we've been recommending to use the separate script
"scripts/build" to build an OSv complete with an application,
e.g., "scripts/build image=rogue".
But for backward compatibility, we also want to support a command like
"make image=rogue" which will then call scripts/build as above.

The way we did this before this patch was somewhat inefficient: it
basically does a "make" first (because of the dependency on "all")
and then runs scripts/build, which ends up calling "make" again.
This duplication becomes even more annoying with uncommitted patches
for ramfs-only images - where this overhead means the image is built
twice (costing about 3 seconds in build time).

So in this patch we do something different: When the "image" option
(for example) is passed, we simply add a new target "default_target"
which only calls scripts/build. Because this target is first in the
Makefile, it is what runs. When image= or modules= are not specified,
these ifdef's do not add these new target, and the first target in
our Makefile returns to being "all", so that is what gets done.

The resulting code is simple and natural, because it plugs into
existing "ifdef"s we already had, and when neither image nor modules
is defined, it does nothing.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1463354799-23737-1-git-send-email-nyh@scylladb.com>