Split building loader.img into 2 stages
1. Build everything needed for loader.elf, except bootfs.o
2. Build the bootfs.o, and link everything to the final loader.elf and
   produce the final loader.img

This is needed for building ramfs-only images (in the next patches).
The scripts/build needs files produced by make. For example
modules/httpserver-jolokia-plugin depends on libtools.so. On the other
hand, scripts/module.py produces binaries and manifests required to
generate bootfs.o.

To brake this recursion, this patch introduces an intermediate target
"stage1" which builds every app dependency we can think of. So
scripts/build is running make twice: before building all applications as
"make stage1", and after - as "make".

Signed-off-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Message-Id: <1463483220-9779-1-git-send-email-iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>