build: move libtools.so to separate module
libtools.so is a OSv-specific library we wrote for use in our "ifconfig" and
"route" implementations (and those REST API operations). Before this patch,
it was compiled as the main OSv Makefile, although it was never actually
needed for building an OSv kernel, but rather just for specific modules
(httpserver and mgmt).

This was always ugly (and we had a comment asking to change this), but it
became a real bug (issue #748) since commit 477035dc2b34, where scripts/build
does not do a full "make" before building the requested modules - it just
does a "make stage1", so libtools.so was not built before the modules which
needed it, so for example a "make clean; make image=httpserver" failed.

The fix in this patch is to move the compilation of libtools.so (and
ifconfig.so and lsroute.so) into the separate module, modules/libtools,
which already existed. The source code of these files is also moved into
modules/libtools. The httpserver module "requires" the libtools module,
so it will get built before httpserver, fixing this bug.

Fixes #748.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1463858848-2973-1-git-send-email-nyh@scylladb.com>