modules/openssl: don't add duplicate libraries
In Linux, shared libraries normally have a short name (like libssl.so.10)
linking to a more specific version of the library (like libssl.so.1.0.2k).
The modules/openssl package also had both names for each library, but
instead of using symbolic links, had two copies of the same library.
This is not a great idea, because some of these libraries are pretty big
(e.g., libcrypto is 2 MB).

In this patch, we completely drop the long names, because the applications
are always linked to (and therefore would require at run time) the short
name. Images which use the openssl module become smaller, for example the
default "scripts/build" image (with cli and rest API) goes down from
27 MB to 25 MB.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1463987740-17232-1-git-send-email-nyh@scylladb.com>