Return error, do not crash, on unknown syscall()
Currently we abort() OSv in case of an unknown syscall() used.
This is too harsh: Linux returns -1 and ENOSYS in that case, and some
applications (such as ElasticSearch) use this to check at runtime
for availability of system calls which not all kernels support.

So this patch replaces the abort() with a warning message and a
return of -1 and errno=ENOSYS.

Fixes #757

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1465118852-27742-1-git-send-email-nyh@scylladb.com>