Remove deprecated "mgmt" submodule
The "mgmt" submodule was a shell (called "Crsh") and ssh implementation
written in Jruby. It predated OSv's REST API and the simple shell written
in Lua on top of it, and once those came out, the old "mgmt" shell became
deprecated and un-recommended.

Moreover, running "scripts/build image=mgmt" hasn't been working for about
a year, because it relies on some external repository holding an
OSv-specific version of the Crsh shell, which is no longer available.

So it's time we get rid of this "mgmt" submodule. Its git repository will
remain (https://github.com/cloudius-systems/mgmt) but it will no longer be
attached to the OSv repository as a submodule.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1467017504-13732-1-git-send-email-nyh@scylladb.com>