scripts/setup.py: fix dependency list for Fedora 23
Fedora 23 recently obsoleted the "python-requests" package in favor of
"python2-request". So now setup.py fails when it tries to install the
obsolete package, when the new one is already installed.

So we need the script to ask for the new name on Fedora 23.
This assumes that people will use the script on an updated system. I think
this is a fair assumption.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>