sched: fix key type of thread_map
Justin Cinkelj <justin.cinkelj@xlab.si> noticed that there is no reason for
the key of "thread_map" to be unsigned long, while our actual thread::id()
is unsigned int.

thread_map's key was a leftover from when thread::id() us indeed an 8-byte
unsigned long. Using 8 bytes for thread ids meant not having to worry about
their wrap-around, but it also caused problems for compatibility with Linux
functions which take a 4-byte thread or process id, so we switched to 4
bytes, but forgot we could make thread_map's key smaller. So this patch
makes it smaller. Actually, instead of assuming which type thread::id()
has, we just tell the compiler to use thread::id()'s return type as the
thread_map key type.

Additionally, this patch replaces an out-of-date comment about why
thread_map is not a set (thread_map is no longer just a list of threads,
but also used to find a thread by its numeric id, so it must be a map and
can't be a set), and makes the thread map and its mutex "static" so they
are only visible in core/sched.cc.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1467531069-20765-1-git-send-email-nyh@scylladb.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>