libc: stub implementation of register_printf_*() functions
glibc has (see http://udrepper.livejournal.com/20948.html) hooks for adding
to printf() support for new types, format specifiers and modifiers. These
functions are used, for example, by the libquadmath library (for 128-bit
quadruple-precision floating point), to add the ability to print 128-bit
floating point numbers.

This patch does not fully implement these hooks - doing this will require
extensive modifications to our printf() implementation, which comes from
Musl. Rather it is just a stub, which prints a warning (once) and returns
an error. Hopefully, this level of support would be enough for applications
which want to use libquadmath (for example) for calculations but do not
intend to use its printf extensions.

Refs issue #765 (this issue should not be closed until we have a full
implementation of these functions, or make a decision that we never will).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1468744037-7194-1-git-send-email-nyh@scylladb.com>