Fix error in xen ring.h
The "&" character was strangely missing in the original code in the
familiar idiom for calculating the offset of a field in a structure.
I don't know how this code ever worked before, but it compiled and
nobody noticed.

In gcc 6, this error caused the compiler to complain about an illegal
conversion of an integer to a pointer (!?) in a huge macro expansion
resulting in a line with over 130,000 characters (!). It wasn't easy
to find the needle in this haystack, but eventually I found it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469049120-7330-1-git-send-email-nyh@scylladb.com>