Use function overload, not template instantiation
Gcc 6 complains when a static template instance is created, but never used -
just like it complains about regular static functions which are not used.

There was no reason in the first place for these functions to be a template.
They can just be run-of-the-mill overloading of the same function names
with different arguments.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469050329-9553-1-git-send-email-nyh@scylladb.com>