pthread: change type of pthread_t
Our pthread_t type was a pointer, while in Linux it was an unsigned long.
This was ok for binary compatibility, because both have the same size,
but the unnecessary difference can cause problems in source code - and
does cause problems for Gcc 6: In Gcc 6's <thread> header file, they
have the code thread::id(1) - which tries to convert the integer 1 into
a pthread_t. This failed when our pthread_t was a pointer.

So this patch trivially changes pthread_t's type to unsigned long, as in
Linux. Nothing else in the code needs to change, because the code always
(see pthread::from_libc) reinterpret_cast's pthread_t, whatever it is,
into a pointer. Additionally, we add a static_assert to verify that these
two types (the pointer and pthread_t) really have the same length, as
intended.

Fixes #770.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469050999-10574-1-git-send-email-nyh@scylladb.com>