sched: ignore thread in wrong wakeup queue
With this patch, if handle_incoming_wakeups() discovers in a CPU's wakeup
queue a thread which does not actually belong to it, it simply ignores it.

This case should never happen. However, in a following patch we add the
ability to pin (migrate) a waiting thread, and that operating might race
with a wake() of the same thread, resulting in the thread remaining on
the old cpu's wakeup queues, and we better ignore it there.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469628899-5537-1-git-send-email-nyh@scylladb.com>