sched: add _pinned flag for threads
Threads currently have a _migration_lock_counter which, when positive,
prevent the thread from being migrated to a different CPU. Using
sched::thread::pin() can make this counter permanently positive, but
we cannot distinguish it from a temporary increase due to migrate_disable().
In the future, to support pinning and re-pinning of other threads, we
want to make such distinction.

So this patch adds a _pinned flag for a thread. When true, the thread was
permanently pinned with pin(). _pinned being set to true also explains an
increase of 1 in _migration_lock_counter.

In the future, the "_pinned" boolean flag should be replaced by a bitmask
of CPUs which this thread is allowed to be on, so that a thread can be
pinned to a set of CPUs instead of just one.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469628899-5537-3-git-send-email-nyh@scylladb.com>