sched: add pin() variant which can pin another thread
We already have sched::thread::pin(cpu*) to migrate the *current* thread to
a given CPU and pin it there. This patch adds a similar function,
pin(thread*, cpu*), to migrate any thread, not just the calling one.

This pin() variant is necessary for a complete implementation of
pthread_setaffinity_np(), for example (see issue #520).

The thread being migrated can of course be running, runnable, or sleeping,
on any of the CPUs, and the code needs to handle all these cases correctly.

The code also needs to handle the case of a thread being temporarily under
migrate_disable(): such a thread must not be migrated, as code surrounded
by migrate_disable() cannot handle being migrated in the middle. So pin()
will sleep a bit, in a loop, until the thread becomes migratable again.
Re-pinning of an already pin()ed thread is supported, though.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469628899-5537-4-git-send-email-nyh@scylladb.com>