sockets: fix recvmsg() function
The recvmsg() function had a bug where the internal implementation modified
the iovecs it receives to keep track of how much it has read. This is
unexpected by callers such as the example in issue #775.

We had exactly the same bug in most of our gather scatter functions
such as readv(), writev(), sendmsg() and fixed them in commits
7b7849c22401675b67b21c96c57955fe99da097f
79c13c3b6bae0172bfb170d2a2bb6356989a50bb
0ce815ad7ca1d117ebe72d66b3b4d3569ac7faa8
respectively. Interestingly while sendmsg() was already fixed, recvmsg()
was not.

We fix the problem here in the same way in all those commits: by making a
temporary copy of the iovec array before calling the internal code which
messes it up.

This patch also adds a test for sendmsg() and recvmsg(), which passes on
Linux, failed on OSv before this patch, and now passes on OSv too.

Fixes #775

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Reviewed-by: Benoit Canet <benoit@scylladb.com>
Message-Id: <1470245226-7199-1-git-send-email-nyh@scylladb.com>