Fix SIOCGIFCONF ioctl with zero size
The SIOCGIFCONF ioctl on sockets is explained in netdevice(7), which says
that if the given ifc_len field is too small to hold the returned data,
the kernel will return partial data and success. Passing ifc_len=0 should
also be ok, although not very useful, but our existing code returned EINVAL.

Musl's if_nameindex() code calls this ioctl with increasing ifc_len starting
with 0, so we need this patch for it not to fail on the first iteration of
its loop.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1470779600-10405-1-git-send-email-nyh@scylladb.com>