Fix if_nameindex()
The if_nameindex() implementation we got from Musl is based on various
ioctls explained in netdevice(7). As that manual page explains, these
ioctls "can be used on any socket's file descriptor regardless of the
family or type.". So it was unfortunate that the musl implementation
chose to create a UNIX domain socket, which is not currently supported
in OSv (see issue #355). Just changing the code to use AF_INET instead
of AF_UNIX makes it work correctly.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1470780495-10772-1-git-send-email-nyh@scylladb.com>