osv::run() do not create a new thread
The osv::run() API waits for the program it started to complete, so it
doesn't really need to start a new thread, and could run the program in
the existing thread.

And yet, it does create a new thread, because it uses the application::run()
API which starts a new application and returns to the caller immediately -
and therefore must create a new thread.

So this patch adds a new method application::run_and_join() which is like
an application::run() followed by an application::join() - except that
control does not return to the caller so we don't need a new thread.
osv::run() then uses this new application::run_and_join().

The osv_execve() suffered from the unnecessary thread in that the actual
application code does not run in the thread id returned by it, but rather
in a child thread. After this patch, it should return the right thread id.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1471867529-2741-1-git-send-email-nyh@scylladb.com>