osv_execve(): fix race in setting returned thread id
osv_execve() promises to return the new thread's id in the given pointer,
but it actually lets the new thread set the id itself - so osv_execve()
may return before the new thread got to run and set this id.

The fix in this patch is for osv_execve() to wait until this variable is
set before returning.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1471903212-13623-1-git-send-email-nyh@scylladb.com>