tests: test for osv::run()
Add a simple test for osv::run. The test tries to run itself, and verifies
that it runs on the same thread (and a new thread is not needlessly
created). It also checks that the return code is returned as expected,
and that running a non-existant object fails with an exception as expected.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1471942378-19972-1-git-send-email-nyh@scylladb.com>