osv::run() - also restore thread's name
application::start_and_join(), and therefore osv::run(), now run the
new application on the same thread. When the application ends, the
thread resumes and we restore the information on which application is
running - but we forgot to also restore the thread's name, which
application::main() overrides. So this patch fixes that.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>