pthread: implement pthread_key_delete()
Implement pthread_key_delete().

Currently, it will not mark the key available for re-use.
The danger in allowing re-use is that if we leave data with that key (the
pthread_key_delete(3P) manpage explains why this happens), we can get a new
destructor running on old unrelated data.

The not-as-great side of not re-using keys is that our current implementation
is limited to 128 keys so we can run out of them. But this patch doesn't make
this any worse than before.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472389371-13580-1-git-send-email-nyh@scylladb.com>
Reviewed-by: Benoît Canet <benoit@scylladb.com>