syscall: set syscall_entry() symbol size to non-zero
With "readelf --dyn-syms" the syscall_entry symbol appears to have a zero
size. This causes our print_backtrace() code to not display the name
"syscall_entry" in the backtrace because no code appears to be inside this
symbol.

We probably need to do this for other assembly functions as well, but this
one bothered me.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472471544-26676-1-git-send-email-nyh@scylladb.com>