syscall: fix CFI information for debugging with GDB
GDB uses DWARF CFI meta-instructions to know how to backtrace through
function call frames, and which registers get saved where.

This patch fixes the CFI instructions in the syscall_entry() assembly
code. After this patch, if we are in system call code we can backtrack
through the syscall_entry() - for example:

  5  0x000000000059c10b in syscall_wrapper (number=186) at linux.cc:332
  6  0x0000000000480245 in syscall_entry () at arch/x64/entry.S:238
  7  0x0000100000c00bce in main (argc=1, argv=<optimized out>)
     at /home/nyh/osv/tests/tst-syscall.cc:49
  8  0x000000000063526e in osv::application::run_main (this=0xffffa00003079910,
     path="tests/tst-syscall.so", argc=1, argv=0xffffa00001f88170)
     at core/app.cc:338

Note how syscall_entry() looks like a normal function - it no longer
pretends to be a "signal frame". In particular, we no longer need to
obey a specific layout of the registers saved on the stack, so I took
this opportunity to remove some of the useless things we saved (like rax,
which will be overwritten by a return value anyway) or saved multiple times
(rcx, rsp).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472503638-5362-2-git-send-email-nyh@scylladb.com>